+++ src/cli/Command.cpp
@@ -23,18 +23,21 @@
 #include "Command.h"
 
 #include "Add.h"
+#include "AddGroup.h"
 #include "Analyze.h"
 #include "Clip.h"
 #include "Create.h"
 #include "Diceware.h"
 #include "Edit.h"
 #include "Estimate.h"
-#include "Extract.h"
+#include "Export.h"
 #include "Generate.h"
 #include "List.h"
 #include "Locate.h"
 #include "Merge.h"
+#include "Move.h"
 #include "Remove.h"
+#include "RemoveGroup.h"
 #include "Show.h"
 #include "TextStream.h"
 #include "Utils.h"
@@ -51,6 +54,12 @@ const QCommandLineOption Command::KeyFileOption = QCommandLineOption(QStringList
 
 const QCommandLineOption Command::NoPasswordOption =
     QCommandLineOption(QStringList() << "no-password", QObject::tr("Deactivate password key for the database."));
+
+const QCommandLineOption Command::YubiKeyOption =
+    QCommandLineOption(QStringList() << "y"
+                                     << "yubikey",
+                       QObject::tr("Yubikey slot used to encrypt the database."),
+                       QObject::tr("slot"));
 
 QMap<QString, Command*> commands;
 
@@ -114,12 +123,15 @@ void populateCommands()
         commands.insert(QString("diceware"), new Diceware());
         commands.insert(QString("edit"), new Edit());
         commands.insert(QString("estimate"), new Estimate());
-        commands.insert(QString("extract"), new Extract());
+        commands.insert(QString("export"), new Export());
         commands.insert(QString("generate"), new Generate());
         commands.insert(QString("locate"), new Locate());
         commands.insert(QString("ls"), new List());
         commands.insert(QString("merge"), new Merge());
+        commands.insert(QString("mkdir"), new AddGroup());
+        commands.insert(QString("mv"), new Move());
         commands.insert(QString("rm"), new Remove());
+        commands.insert(QString("rmdir"), new RemoveGroup());
         commands.insert(QString("show"), new Show());
     }
 }