+++ src/cli/Move.h
@@ -15,23 +15,18 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef KEEPASSXC_MERGE_H
-#define KEEPASSXC_MERGE_H
+#ifndef KEEPASSXC_MOVE_H
+#define KEEPASSXC_MOVE_H
 
 #include "DatabaseCommand.h"
 
-class Merge : public DatabaseCommand
+class Move : public DatabaseCommand
 {
 public:
-    Merge();
+    Move();
+    ~Move();
 
-    int executeWithDatabase(QSharedPointer<Database> db, QSharedPointer<QCommandLineParser> parser) override;
-
-    static const QCommandLineOption SameCredentialsOption;
-    static const QCommandLineOption KeyFileFromOption;
-    static const QCommandLineOption NoPasswordFromOption;
-    static const QCommandLineOption YubiKeyFromOption;
-    static const QCommandLineOption DryRunOption;
+    int executeWithDatabase(QSharedPointer<Database> db, QSharedPointer<QCommandLineParser> parser);
 };
 
-#endif // KEEPASSXC_MERGE_H
+#endif // KEEPASSXC_MOVE_H