+++ tests/TestCli.cpp
@@ -32,6 +32,7 @@
 #include "format/KeePass2.h"
 
 #include "cli/Add.h"
+#include "cli/AddGroup.h"
 #include "cli/Analyze.h"
 #include "cli/Clip.h"
 #include "cli/Command.h"
@@ -39,12 +40,14 @@
 #include "cli/Diceware.h"
 #include "cli/Edit.h"
 #include "cli/Estimate.h"
-#include "cli/Extract.h"
+#include "cli/Export.h"
 #include "cli/Generate.h"
 #include "cli/List.h"
 #include "cli/Locate.h"
 #include "cli/Merge.h"
+#include "cli/Move.h"
 #include "cli/Remove.h"
+#include "cli/RemoveGroup.h"
 #include "cli/Show.h"
 #include "cli/Utils.h"
 
@@ -89,6 +92,11 @@ void TestCli::initTestCase()
     QVERIFY(sourceDbFile4.open(QIODevice::ReadOnly));
     QVERIFY(Tools::readAllFromDevice(&sourceDbFile4, m_keyFileProtectedNoPasswordDbData));
     sourceDbFile4.close();
+
+    QFile sourceDbFileYubiKeyProtected(QString(KEEPASSX_TEST_DATA_DIR).append("/YubiKeyProtectedPasswords.kdbx"));
+    QVERIFY(sourceDbFileYubiKeyProtected.open(QIODevice::ReadOnly));
+    QVERIFY(Tools::readAllFromDevice(&sourceDbFileYubiKeyProtected, m_yubiKeyProtectedDbData));
+    sourceDbFileYubiKeyProtected.close();
 }
 
 void TestCli::init()
@@ -112,6 +120,11 @@ void TestCli::init()
     m_keyFileProtectedNoPasswordDbFile->open();
     m_keyFileProtectedNoPasswordDbFile->write(m_keyFileProtectedNoPasswordDbData);
     m_keyFileProtectedNoPasswordDbFile->close();
+
+    m_yubiKeyProtectedDbFile.reset(new TemporaryFile());
+    m_yubiKeyProtectedDbFile->open();
+    m_yubiKeyProtectedDbFile->write(m_yubiKeyProtectedDbData);
+    m_yubiKeyProtectedDbFile->close();
 
     m_stdinFile.reset(new TemporaryFile());
     m_stdinFile->open();
@@ -155,14 +168,14 @@ void TestCli::cleanupTestCase()
 QSharedPointer<Database> TestCli::readTestDatabase() const
 {
     Utils::Test::setNextPassword("a");
-    auto db = QSharedPointer<Database>(Utils::unlockDatabase(m_dbFile->fileName(), true, "", m_stdoutHandle));
+    auto db = QSharedPointer<Database>(Utils::unlockDatabase(m_dbFile->fileName(), true, "", "", m_stdoutHandle));
     m_stdoutFile->seek(ftell(m_stdoutHandle)); // re-synchronize handles
     return db;
 }
 
 void TestCli::testCommand()
 {
-    QCOMPARE(Command::getCommands().size(), 14);
+    QCOMPARE(Command::getCommands().size(), 17);
     QVERIFY(Command::getCommand("add"));
     QVERIFY(Command::getCommand("analyze"));
     QVERIFY(Command::getCommand("clip"));
@@ -170,7 +183,7 @@ void TestCli::testCommand()
     QVERIFY(Command::getCommand("diceware"));
     QVERIFY(Command::getCommand("edit"));
     QVERIFY(Command::getCommand("estimate"));
-    QVERIFY(Command::getCommand("extract"));
+    QVERIFY(Command::getCommand("export"));
     QVERIFY(Command::getCommand("generate"));
     QVERIFY(Command::getCommand("locate"));
     QVERIFY(Command::getCommand("ls"));
@@ -226,14 +239,8 @@ void TestCli::testAdd()
 
     Utils::Test::setNextPassword("a");
     Utils::Test::setNextPassword("newpassword");
-    addCmd.execute({"add",
-                    "-u",
-                    "newuser2",
-                    "--url",
-                    "https://example.net/",
-                    "-p",
-                    m_dbFile->fileName(),
-                    "/newuser-entry2"});
+    addCmd.execute(
+        {"add", "-u", "newuser2", "--url", "https://example.net/", "-p", m_dbFile->fileName(), "/newuser-entry2"});
 
     db = readTestDatabase();
     entry = db->rootGroup()->findEntryByPath("/newuser-entry2");
@@ -246,14 +253,7 @@ void TestCli::testAdd()
     pos = m_stdoutFile->pos();
     posErr = m_stderrFile->pos();
     Utils::Test::setNextPassword("a");
-    addCmd.execute({"add",
-                    "-u",
-                    "newuser3",
-                    "-g",
-                    "-L",
-                    "34",
-                    m_dbFile->fileName(),
-                    "/newuser-entry3"});
+    addCmd.execute({"add", "-u", "newuser3", "-g", "-L", "34", m_dbFile->fileName(), "/newuser-entry3"});
     m_stdoutFile->seek(pos);
     m_stderrFile->seek(posErr);
     m_stdoutFile->readLine(); // skip password prompt
@@ -298,20 +298,73 @@ void TestCli::testAdd()
     QVERIFY(!defaultPasswordClassesRegex.match(entry->password()).hasMatch());
 }
 
-void TestCli::testAnalyze()
+void TestCli::testAddGroup()
 {
-    Analyze analyzeCmd;
-    QVERIFY(!analyzeCmd.name.isEmpty());
-    QVERIFY(analyzeCmd.getDescriptionLine().contains(analyzeCmd.name));
-
-    const QString hibpPath = QString(KEEPASSX_TEST_DATA_DIR).append("/hibp.txt");
+    AddGroup addGroupCmd;
+    QVERIFY(!addGroupCmd.name.isEmpty());
+    QVERIFY(addGroupCmd.getDescriptionLine().contains(addGroupCmd.name));
 
     Utils::Test::setNextPassword("a");
-    analyzeCmd.execute({"analyze", "--hibp", hibpPath, m_dbFile->fileName()});
+    addGroupCmd.execute({"mkdir", m_dbFile->fileName(), "/new_group"});
+    m_stderrFile->reset();
     m_stdoutFile->reset();
     m_stdoutFile->readLine(); // skip password prompt
-    auto output = m_stdoutFile->readAll();
-    QVERIFY(output.contains("Sample Entry") && output.contains("123"));
+    QCOMPARE(m_stderrFile->readAll(), QByteArray(""));
+    QCOMPARE(m_stdoutFile->readAll(), QByteArray("Successfully added group new_group.\n"));
+
+    auto db = readTestDatabase();
+    auto* group = db->rootGroup()->findGroupByPath("new_group");
+    QVERIFY(group);
+    QCOMPARE(group->name(), QString("new_group"));
+
+    // Trying to add the same group should fail.
+    qint64 pos = m_stdoutFile->pos();
+    qint64 posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    addGroupCmd.execute({"mkdir", m_dbFile->fileName(), "/new_group"});
+    m_stdoutFile->seek(pos);
+    m_stdoutFile->readLine(); // skip password prompt
+    m_stderrFile->seek(posErr);
+    QCOMPARE(m_stderrFile->readAll(), QByteArray("Group /new_group already exists!\n"));
+    QCOMPARE(m_stdoutFile->readAll(), QByteArray(""));
+
+    // Should be able to add groups down the tree.
+    pos = m_stdoutFile->pos();
+    posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    addGroupCmd.execute({"mkdir", m_dbFile->fileName(), "/new_group/newer_group"});
+    m_stdoutFile->seek(pos);
+    m_stdoutFile->readLine(); // skip password prompt
+    m_stderrFile->seek(posErr);
+    QCOMPARE(m_stderrFile->readAll(), QByteArray(""));
+    QCOMPARE(m_stdoutFile->readAll(), QByteArray("Successfully added group newer_group.\n"));
+
+    db = readTestDatabase();
+    group = db->rootGroup()->findGroupByPath("new_group/newer_group");
+    QVERIFY(group);
+    QCOMPARE(group->name(), QString("newer_group"));
+
+    // Should fail if the path is invalid.
+    pos = m_stdoutFile->pos();
+    posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    addGroupCmd.execute({"mkdir", m_dbFile->fileName(), "/invalid_group/newer_group"});
+    m_stdoutFile->seek(pos);
+    m_stdoutFile->readLine(); // skip password prompt
+    m_stderrFile->seek(posErr);
+    QCOMPARE(m_stderrFile->readAll(), QByteArray("Group /invalid_group not found.\n"));
+    QCOMPARE(m_stdoutFile->readAll(), QByteArray(""));
+
+    // Should fail to add the root group.
+    pos = m_stdoutFile->pos();
+    posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    addGroupCmd.execute({"mkdir", m_dbFile->fileName(), "/"});
+    m_stdoutFile->seek(pos);
+    m_stdoutFile->readLine(); // skip password prompt
+    m_stderrFile->seek(posErr);
+    QCOMPARE(m_stderrFile->readAll(), QByteArray("Group / already exists!\n"));
+    QCOMPARE(m_stdoutFile->readAll(), QByteArray(""));
 }
 
 bool isTOTP(const QString& value)
@@ -326,6 +379,22 @@ bool isTOTP(const QString& value)
         }
     }
     return true;
+}
+
+void TestCli::testAnalyze()
+{
+    Analyze analyzeCmd;
+    QVERIFY(!analyzeCmd.name.isEmpty());
+    QVERIFY(analyzeCmd.getDescriptionLine().contains(analyzeCmd.name));
+
+    const QString hibpPath = QString(KEEPASSX_TEST_DATA_DIR).append("/hibp.txt");
+
+    Utils::Test::setNextPassword("a");
+    analyzeCmd.execute({"analyze", "--hibp", hibpPath, m_dbFile->fileName()});
+    m_stdoutFile->reset();
+    m_stdoutFile->readLine(); // skip password prompt
+    auto output = m_stdoutFile->readAll();
+    QVERIFY(output.contains("Sample Entry") && output.contains("123"));
 }
 
 void TestCli::testClip()
@@ -433,7 +502,7 @@ void TestCli::testCreate()
     QCOMPARE(m_stdoutFile->readLine(), QByteArray("Successfully created new database.\n"));
 
     Utils::Test::setNextPassword("a");
-    auto db = QSharedPointer<Database>(Utils::unlockDatabase(databaseFilename, true, "", Utils::DEVNULL));
+    auto db = QSharedPointer<Database>(Utils::unlockDatabase(databaseFilename, true, "", "", Utils::DEVNULL));
     QVERIFY(db);
 
     // Should refuse to create the database if it already exists.
@@ -462,7 +531,8 @@ void TestCli::testCreate()
     QCOMPARE(m_stdoutFile->readLine(), QByteArray("Successfully created new database.\n"));
 
     Utils::Test::setNextPassword("a");
-    auto db2 = QSharedPointer<Database>(Utils::unlockDatabase(databaseFilename2, true, keyfilePath, Utils::DEVNULL));
+    auto db2 =
+        QSharedPointer<Database>(Utils::unlockDatabase(databaseFilename2, true, keyfilePath, "", Utils::DEVNULL));
     QVERIFY(db2);
 
     // Testing with existing keyfile
@@ -479,7 +549,8 @@ void TestCli::testCreate()
     QCOMPARE(m_stdoutFile->readLine(), QByteArray("Successfully created new database.\n"));
 
     Utils::Test::setNextPassword("a");
-    auto db3 = QSharedPointer<Database>(Utils::unlockDatabase(databaseFilename3, true, keyfilePath, Utils::DEVNULL));
+    auto db3 =
+        QSharedPointer<Database>(Utils::unlockDatabase(databaseFilename3, true, keyfilePath, "", Utils::DEVNULL));
     QVERIFY(db3);
 }
 
@@ -742,14 +813,14 @@ void TestCli::testEstimate()
     }
 }
 
-void TestCli::testExtract()
+void TestCli::testExport()
 {
-    Extract extractCmd;
-    QVERIFY(!extractCmd.name.isEmpty());
-    QVERIFY(extractCmd.getDescriptionLine().contains(extractCmd.name));
+    Export exportCmd;
+    QVERIFY(!exportCmd.name.isEmpty());
+    QVERIFY(exportCmd.getDescriptionLine().contains(exportCmd.name));
 
     Utils::Test::setNextPassword("a");
-    extractCmd.execute({"extract", m_dbFile->fileName()});
+    exportCmd.execute({"export", m_dbFile->fileName()});
 
     m_stdoutFile->seek(0);
     m_stdoutFile->readLine(); // skip prompt line
@@ -768,12 +839,41 @@ void TestCli::testExtract()
     // Quiet option
     QScopedPointer<Database> dbQuiet(new Database());
     qint64 pos = m_stdoutFile->pos();
+    qint64 posErr = m_stderrFile->pos();
     Utils::Test::setNextPassword("a");
-    extractCmd.execute({"extract", "-q", m_dbFile->fileName()});
+    exportCmd.execute({"export", "-f", "xml", "-q", m_dbFile->fileName()});
     m_stdoutFile->seek(pos);
+    m_stderrFile->seek(posErr);
     reader.readDatabase(m_stdoutFile.data(), dbQuiet.data());
     QVERIFY(!reader.hasError());
     QVERIFY(db.data());
+    QCOMPARE(m_stderrFile->readAll(), QByteArray(""));
+
+    // CSV exporting
+    pos = m_stdoutFile->pos();
+    posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    exportCmd.execute({"export", "-f", "csv", m_dbFile->fileName()});
+    m_stdoutFile->seek(pos);
+    m_stdoutFile->readLine(); // skip prompt line
+    m_stderrFile->seek(posErr);
+    QByteArray csvHeader = m_stdoutFile->readLine();
+    QCOMPARE(csvHeader, QByteArray("\"Group\",\"Title\",\"Username\",\"Password\",\"URL\",\"Notes\"\n"));
+    QByteArray csvData = m_stdoutFile->readAll();
+    QVERIFY(csvData.contains(QByteArray(
+        "\"NewDatabase\",\"Sample Entry\",\"User Name\",\"Password\",\"http://www.somesite.com/\",\"Notes\"\n")));
+    QCOMPARE(m_stderrFile->readAll(), QByteArray(""));
+
+    // test invalid format
+    pos = m_stdoutFile->pos();
+    posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    exportCmd.execute({"export", "-f", "yaml", m_dbFile->fileName()});
+    m_stdoutFile->seek(pos);
+    m_stdoutFile->readLine(); // skip prompt line
+    m_stderrFile->seek(posErr);
+    QCOMPARE(m_stdoutFile->readLine(), QByteArray(""));
+    QCOMPARE(m_stderrFile->readLine(), QByteArray("Unsupported format yaml\n"));
 }
 
 void TestCli::testGenerate_data()
@@ -1217,6 +1317,63 @@ void TestCli::testMerge()
     QCOMPARE(m_stdoutFile->readAll(), QByteArray(""));
 }
 
+void TestCli::testMove()
+{
+    Move moveCmd;
+    QVERIFY(!moveCmd.name.isEmpty());
+    QVERIFY(moveCmd.getDescriptionLine().contains(moveCmd.name));
+
+    qint64 pos = m_stdoutFile->pos();
+    qint64 posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    moveCmd.execute({"mv", m_dbFile->fileName(), "invalid_entry_path", "invalid_group_path"});
+    m_stdoutFile->seek(pos);
+    m_stderrFile->seek(posErr);
+    m_stdoutFile->readLine(); // skip prompt line
+    QCOMPARE(m_stdoutFile->readLine(), QByteArray(""));
+    QCOMPARE(m_stderrFile->readLine(), QByteArray("Could not find entry with path invalid_entry_path.\n"));
+
+    pos = m_stdoutFile->pos();
+    posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    moveCmd.execute({"mv", m_dbFile->fileName(), "Sample Entry", "invalid_group_path"});
+    m_stdoutFile->seek(pos);
+    m_stderrFile->seek(posErr);
+    m_stdoutFile->readLine(); // skip prompt line
+    QCOMPARE(m_stdoutFile->readLine(), QByteArray(""));
+    QCOMPARE(m_stderrFile->readLine(), QByteArray("Could not find group with path invalid_group_path.\n"));
+
+    pos = m_stdoutFile->pos();
+    posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    moveCmd.execute({"mv", m_dbFile->fileName(), "Sample Entry", "General/"});
+    m_stdoutFile->seek(pos);
+    m_stderrFile->seek(posErr);
+    m_stdoutFile->readLine(); // skip prompt line
+    QCOMPARE(m_stdoutFile->readLine(), QByteArray("Successfully moved entry Sample Entry to group General/.\n"));
+    QCOMPARE(m_stderrFile->readLine(), QByteArray(""));
+
+    auto db = readTestDatabase();
+    auto* entry = db->rootGroup()->findEntryByPath("General/Sample Entry");
+    QVERIFY(entry);
+
+    // Test that not modified if the same group is destination.
+    pos = m_stdoutFile->pos();
+    posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    moveCmd.execute({"mv", m_dbFile->fileName(), "General/Sample Entry", "General/"});
+    m_stdoutFile->seek(pos);
+    m_stderrFile->seek(posErr);
+    m_stdoutFile->readLine(); // skip prompt line
+    QCOMPARE(m_stdoutFile->readLine(), QByteArray(""));
+    QCOMPARE(m_stderrFile->readLine(), QByteArray("Entry is already in group General/.\n"));
+
+    // sanity check
+    db = readTestDatabase();
+    entry = db->rootGroup()->findEntryByPath("General/Sample Entry");
+    QVERIFY(entry);
+}
+
 void TestCli::testRemove()
 {
     Remove removeCmd;
@@ -1236,6 +1393,7 @@ void TestCli::testRemove()
     fileCopy.close();
 
     qint64 pos = m_stdoutFile->pos();
+    qint64 posErr = m_stderrFile->pos();
 
     // delete entry and verify
     Utils::Test::setNextPassword("a");
@@ -1243,6 +1401,7 @@ void TestCli::testRemove()
     m_stdoutFile->seek(pos);
     m_stdoutFile->readLine(); // skip password prompt
     QCOMPARE(m_stdoutFile->readAll(), QByteArray("Successfully recycled entry Sample Entry.\n"));
+    QCOMPARE(m_stderrFile->readAll(), QByteArray(""));
 
     auto key = QSharedPointer<CompositeKey>::create();
     key->addKey(QSharedPointer<PasswordKey>::create("a"));
@@ -1255,6 +1414,7 @@ void TestCli::testRemove()
     QVERIFY(!readBackDb->rootGroup()->findEntryByPath("/Sample Entry"));
     QVERIFY(readBackDb->rootGroup()->findEntryByPath(QString("/%1/Sample Entry").arg(Group::tr("Recycle Bin"))));
 
+    pos = m_stdoutFile->pos();
     pos = m_stdoutFile->pos();
 
     // try again, this time without recycle bin
@@ -1273,16 +1433,89 @@ void TestCli::testRemove()
     QVERIFY(!readBackDb->rootGroup()->findEntryByPath("/Sample Entry"));
     QVERIFY(!readBackDb->rootGroup()->findEntryByPath(QString("/%1/Sample Entry").arg(Group::tr("Recycle Bin"))));
 
-    pos = m_stdoutFile->pos();
-
     // finally, try deleting a non-existent entry
+    pos = m_stdoutFile->pos();
+    posErr = m_stderrFile->pos();
     Utils::Test::setNextPassword("a");
     removeCmd.execute({"rm", fileCopy.fileName(), "/Sample Entry"});
     m_stdoutFile->seek(pos);
     m_stdoutFile->readLine(); // skip password prompt
-    m_stderrFile->reset();
+    m_stderrFile->seek(posErr);
     QCOMPARE(m_stdoutFile->readAll(), QByteArray(""));
     QCOMPARE(m_stderrFile->readAll(), QByteArray("Entry /Sample Entry not found.\n"));
+
+    // try deleting a directory, should fail
+    pos = m_stdoutFile->pos();
+    posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    removeCmd.execute({"rm", fileCopy.fileName(), "/General"});
+    m_stdoutFile->seek(pos);
+    m_stdoutFile->readLine(); // skip password prompt
+    m_stderrFile->seek(posErr);
+    QCOMPARE(m_stdoutFile->readAll(), QByteArray(""));
+    QCOMPARE(m_stderrFile->readAll(), QByteArray("Entry /General not found.\n"));
+}
+
+void TestCli::testRemoveGroup()
+{
+    RemoveGroup removeGroupCmd;
+    QVERIFY(!removeGroupCmd.name.isEmpty());
+    QVERIFY(removeGroupCmd.getDescriptionLine().contains(removeGroupCmd.name));
+
+    Kdbx3Reader reader;
+    Kdbx3Writer writer;
+
+    // try deleting a directory, should recycle it first.
+    qint64 pos = m_stdoutFile->pos();
+    qint64 posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    removeGroupCmd.execute({"rmdir", m_dbFile->fileName(), "/General"});
+    m_stdoutFile->seek(pos);
+    m_stdoutFile->readLine(); // skip password prompt
+    m_stderrFile->seek(posErr);
+    QCOMPARE(m_stderrFile->readAll(), QByteArray(""));
+    QCOMPARE(m_stdoutFile->readAll(), QByteArray("Successfully recycled group /General.\n"));
+
+    auto db = readTestDatabase();
+    auto* group = db->rootGroup()->findGroupByPath("General");
+    QVERIFY(!group);
+
+    // try deleting a directory again, should delete it permanently.
+    pos = m_stdoutFile->pos();
+    posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    removeGroupCmd.execute({"rmdir", m_dbFile->fileName(), "Recycle Bin/General"});
+    m_stdoutFile->seek(pos);
+    m_stdoutFile->readLine(); // skip password prompt
+    m_stderrFile->seek(posErr);
+    QCOMPARE(m_stdoutFile->readAll(), QByteArray("Successfully deleted group Recycle Bin/General.\n"));
+    QCOMPARE(m_stderrFile->readAll(), QByteArray(""));
+
+    db = readTestDatabase();
+    group = db->rootGroup()->findGroupByPath("Recycle Bin/General");
+    QVERIFY(!group);
+
+    // try deleting an invalid group, should fail.
+    pos = m_stdoutFile->pos();
+    posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    removeGroupCmd.execute({"rmdir", m_dbFile->fileName(), "invalid"});
+    m_stdoutFile->seek(pos);
+    m_stdoutFile->readLine(); // skip password prompt
+    m_stderrFile->seek(posErr);
+    QCOMPARE(m_stderrFile->readAll(), QByteArray("Group invalid not found.\n"));
+    QCOMPARE(m_stdoutFile->readAll(), QByteArray(""));
+
+    // Should fail to remove the root group.
+    pos = m_stdoutFile->pos();
+    posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    removeGroupCmd.execute({"rmdir", m_dbFile->fileName(), "/"});
+    m_stdoutFile->seek(pos);
+    m_stdoutFile->readLine(); // skip password prompt
+    m_stderrFile->seek(posErr);
+    QCOMPARE(m_stderrFile->readAll(), QByteArray("Cannot remove root group from database.\n"));
+    QCOMPARE(m_stdoutFile->readAll(), QByteArray(""));
 }
 
 void TestCli::testRemoveQuiet()
@@ -1408,4 +1641,112 @@ void TestCli::testShow()
     m_stderrFile->seek(posErr);
     QCOMPARE(m_stdoutFile->readAll(), QByteArray(""));
     QCOMPARE(m_stderrFile->readAll(), QByteArray("Entry with path /Sample Entry has no TOTP set up.\n"));
+}
+
+void TestCli::testInvalidDbFiles()
+{
+    Show showCmd;
+    QString nonExistentDbPath("/foo/bar/baz");
+    QString directoryName("/");
+
+    qint64 pos = m_stderrFile->pos();
+    showCmd.execute({"show", nonExistentDbPath, "-q", "/Sample Entry"});
+    m_stderrFile->seek(pos);
+    QCOMPARE(QString(m_stderrFile->readAll()),
+             QObject::tr("Failed to open database file %1: not found").arg(nonExistentDbPath) + "\n");
+
+    pos = m_stderrFile->pos();
+    showCmd.execute({"show", directoryName, "-q", "whatever"});
+    m_stderrFile->seek(pos);
+    QCOMPARE(QString(m_stderrFile->readAll()),
+             QObject::tr("Failed to open database file %1: not a plain file").arg(directoryName) + "\n");
+
+    // Create a write-only file and try to open it.
+    // QFileInfo.isReadable returns 'true' on Windows, even after the call to
+    // setPermissions(WriteOwner) and with NTFS permissions enabled, so this
+    // check doesn't work.
+#if !defined(Q_OS_WIN)
+    QTemporaryFile tempFile;
+    QVERIFY(tempFile.open());
+    QString path = QFileInfo(tempFile).absoluteFilePath();
+    QVERIFY(tempFile.setPermissions(QFileDevice::WriteOwner));
+    pos = m_stderrFile->pos();
+    showCmd.execute({"show", path, "some entry"});
+    m_stderrFile->seek(pos);
+    QCOMPARE(QString(m_stderrFile->readAll()),
+             QObject::tr("Failed to open database file %1: not readable").arg(path) + "\n");
+#endif // Q_OS_WIN
+}
+
+/**
+ * Secret key for the YubiKey slot used by the unit test is
+ * 1c e3 0f d7 8d 20 dc fa 40 b5 0c 18 77 9a fb 0f 02 28 8d b7
+ * This secret should be configured at slot 2, and the slot
+ * should be configured as passive.
+ */
+void TestCli::testYubiKeyOption()
+{
+    if (!YubiKey::instance()->init()) {
+        QSKIP("Unable to connect to YubiKey");
+    }
+
+    QString errorMessage;
+    bool isBlocking = YubiKey::instance()->checkSlotIsBlocking(2, errorMessage);
+    if (isBlocking && errorMessage.isEmpty()) {
+        QSKIP("Skipping YubiKey in press mode.");
+    }
+
+    QByteArray challenge("CLITest");
+    QByteArray response;
+    YubiKey::instance()->challenge(2, false, challenge, response);
+    QByteArray expected("\xA2\x3B\x94\x00\xBE\x47\x9A\x30\xA9\xEB\x50\x9B\x85\x56\x5B\x6B\x30\x25\xB4\x8E", 20);
+    QVERIFY2(response == expected, "YubiKey Slot 2 is not configured with correct secret key.");
+
+    List listCmd;
+    Add addCmd;
+
+    Utils::Test::setNextPassword("a");
+    listCmd.execute({"ls", "-y", "2", m_yubiKeyProtectedDbFile->fileName()});
+    m_stdoutFile->reset();
+    m_stderrFile->reset();
+    m_stdoutFile->readLine(); // skip password prompt
+    QCOMPARE(m_stdoutFile->readAll(),
+             QByteArray("entry1\n"
+                        "entry2\n"));
+
+    // Should raise an error with no yubikey slot.
+    qint64 pos = m_stdoutFile->pos();
+    qint64 posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    listCmd.execute({"ls", m_yubiKeyProtectedDbFile->fileName()});
+    m_stdoutFile->seek(pos);
+    m_stdoutFile->readLine(); // skip password prompt
+    m_stderrFile->seek(posErr);
+    QCOMPARE(m_stdoutFile->readAll(), QByteArray(""));
+    QCOMPARE(m_stderrFile->readLine(),
+             QByteArray("Error while reading the database: Invalid credentials were provided, please try again.\n"));
+    QCOMPARE(m_stderrFile->readLine(),
+             QByteArray("If this reoccurs, then your database file may be corrupt. (HMAC mismatch)\n"));
+
+    // Should raise an error if yubikey slot is not a string
+    pos = m_stdoutFile->pos();
+    posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    listCmd.execute({"ls", "-y", "invalidslot", m_yubiKeyProtectedDbFile->fileName()});
+    m_stdoutFile->seek(pos);
+    m_stdoutFile->readLine(); // skip password prompt
+    m_stderrFile->seek(posErr);
+    QCOMPARE(m_stdoutFile->readAll(), QByteArray(""));
+    QCOMPARE(m_stderrFile->readAll().split(':').at(0), QByteArray("Invalid YubiKey slot invalidslot\n"));
+
+    // Should raise an error if yubikey slot is invalid.
+    pos = m_stdoutFile->pos();
+    posErr = m_stderrFile->pos();
+    Utils::Test::setNextPassword("a");
+    listCmd.execute({"ls", "-y", "3", m_yubiKeyProtectedDbFile->fileName()});
+    m_stdoutFile->seek(pos);
+    m_stdoutFile->readLine(); // skip password prompt
+    m_stderrFile->seek(posErr);
+    QCOMPARE(m_stdoutFile->readAll(), QByteArray(""));
+    QCOMPARE(m_stderrFile->readAll().split(':').at(0), QByteArray("Invalid YubiKey slot 3\n"));
 }