+++ tests/TestCli.h
@@ -45,6 +45,7 @@ private slots:
 
     void testCommand();
     void testAdd();
+    void testAddGroup();
     void testAnalyze();
     void testClip();
     void testCreate();
@@ -52,7 +53,7 @@ private slots:
     void testEdit();
     void testEstimate_data();
     void testEstimate();
-    void testExtract();
+    void testExport();
     void testGenerate_data();
     void testGenerate();
     void testKeyFileOption();
@@ -60,19 +61,25 @@ private slots:
     void testList();
     void testLocate();
     void testMerge();
+    void testMove();
     void testRemove();
+    void testRemoveGroup();
     void testRemoveQuiet();
     void testShow();
+    void testInvalidDbFiles();
+    void testYubiKeyOption();
 
 private:
     QByteArray m_dbData;
     QByteArray m_dbData2;
+    QByteArray m_yubiKeyProtectedDbData;
     QByteArray m_keyFileProtectedDbData;
     QByteArray m_keyFileProtectedNoPasswordDbData;
     QScopedPointer<TemporaryFile> m_dbFile;
     QScopedPointer<TemporaryFile> m_dbFile2;
     QScopedPointer<TemporaryFile> m_keyFileProtectedDbFile;
     QScopedPointer<TemporaryFile> m_keyFileProtectedNoPasswordDbFile;
+    QScopedPointer<TemporaryFile> m_yubiKeyProtectedDbFile;
     QScopedPointer<TemporaryFile> m_stdoutFile;
     QScopedPointer<TemporaryFile> m_stderrFile;
     QScopedPointer<TemporaryFile> m_stdinFile;